+++ src/number.ts
@@ -1,6 +1,7 @@
 /**
  * @since 2.10.0
  */
+import * as B from './Bounded'
 import * as E from './Eq'
 import * as O from './Ord'
 
@@ -17,4 +18,15 @@ export const Eq: E.Eq<number> = E.eqStrict
 export const Ord: O.Ord<number> = {
   equals: Eq.equals,
   compare: (x, y) => (x < y ? -1 : x > y ? 1 : 0)
+}
+
+/**
+ * @category instances
+ * @since 2.10.0
+ */
+export const Bounded: B.Bounded<number> = {
+  equals: Ord.equals,
+  compare: Ord.compare,
+  top: Infinity,
+  bottom: -Infinity
 }